fix(meta): use object store's timestamp in SST retention sanity check

+++ src/object_store/src/object/mem.rs
@@ -288,7 +288,7 @@ fn get_obj_meta(path: &str, obj: &Bytes) -> ObjectResult<ObjectMetadata> {
         last_modified: SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .map_err(ObjectError::internal)?
-            .as_secs_f64(),
+            .as_secs_f64() as u64,
         total_size: obj.len(),
     })
 }
